backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.26.2
opencv-python==4.8.1.78
scikit-learn==1.3.2
pillow==10.1.0
python-dotenv==1.0.0
joblib==1.3.2
requests==2.31.0
pydantic==2.5.2
